Add new material files, update shaders, and modify .gitignore for ImGui settings
