+++ .gitignore
@@ -4,3 +4,4 @@ target/*
 # src/main/resources/*
 .VSCodeCounter/*
 .idea/
+imgui.ini
