+++ pom.xml
@@ -14,6 +14,7 @@
 
         <lwjgl.version>3.3.6</lwjgl.version>
         <joml.version>1.10.7</joml.version>
+        <imgui.version>1.87.0</imgui.version>
     </properties>
 
     <profiles>
@@ -108,13 +109,15 @@
         </dependency>
 
         <dependency>
-
             <groupId>org.apache.commons</groupId>
-
             <artifactId>commons-lang3</artifactId>
-
             <version>3.12.0</version>
+        </dependency>
 
+        <dependency>
+            <groupId>io.github.spair</groupId>
+            <artifactId>imgui-java-app</artifactId>
+            <version>${imgui.version}</version>
         </dependency>
     </dependencies>
