--- src/main/java/com/steve/graphic/Sphere.java
@@ -1,74 +1,0 @@
-package com.steve.graphic;
-
-public class Sphere {
-    
-    static float[] vertices = new float[] {
-        // positions          // colors           // texture coords
-        0.000000f, -1.000000f,  0.000000f,  1.0f, 1.0f, 1.0f,  0.181819f, 0.000000f,
-        0.723607f, -0.447220f,  0.525725f,  1.0f, 1.0f, 1.0f,  0.227273f, 0.078731f,
-        -0.276388f, -0.447220f,  0.850649f,  1.0f, 1.0f, 1.0f,  0.136365f, 0.078731f,
-        -0.894426f, -0.447216f,  0.000000f,  1.0f, 1.0f, 1.0f,  0.272728f, 0.157461f,
-        -0.276388f, -0.447220f, -0.850649f,  1.0f, 1.0f, 1.0f,  0.318182f, 0.078731f,
-        0.723607f, -0.447220f, -0.525725f,  1.0f, 1.0f, 1.0f,  0.363637f, 0.157461f,
-        0.276388f,  0.447220f,  0.850649f,  1.0f, 1.0f, 1.0f,  0.909091f, 0.000000f,
-        -0.723607f,  0.447220f,  0.525725f,  1.0f, 1.0f, 1.0f,  0.954545f, 0.078731f,
-        -0.723607f,  0.447220f, -0.525725f,  1.0f, 1.0f, 1.0f,  0.863636f, 0.078731f,
-        0.276388f,  0.447220f, -0.850649f,  1.0f, 1.0f, 1.0f,  0.727273f, 0.000000f,
-        0.894426f,  0.447216f,  0.000000f,  1.0f, 1.0f, 1.0f,  0.772727f, 0.078731f,
-        0.000000f,  1.000000f,  0.000000f,  1.0f, 1.0f, 1.0f,  0.681818f, 0.078731f,
-        -0.162456f, -0.850654f,  0.499995f,  1.0f, 1.0f, 1.0f,  0.545455f, 0.000000f,
-        0.425323f, -0.850654f,  0.309011f,  1.0f, 1.0f, 1.0f,  0.590909f, 0.078731f,
-        0.262869f, -0.525738f,  0.809012f,  1.0f, 1.0f, 1.0f,  0.500000f, 0.078731f,
-        0.850648f, -0.525736f,  0.000000f,  1.0f, 1.0f, 1.0f,  0.318182f, 0.236191f,
-        0.425323f, -0.850654f, -0.309011f,  1.0f, 1.0f, 1.0f,  0.090910f, 0.157461f,
-        -0.525730f, -0.850652f,  0.000000f,  1.0f, 1.0f, 1.0f,  0.181819f, 0.157461f,
-        -0.688189f, -0.525736f,  0.499997f,  1.0f, 1.0f, 1.0f,  0.136365f, 0.236191f,
-        -0.162456f, -0.850654f, -0.499995f,  1.0f, 1.0f, 1.0f,  0.818182f, 0.157461f,
-        -0.688189f, -0.525736f, -0.499997f,  1.0f, 1.0f, 1.0f,  0.909091f, 0.157461f,
-        0.262869f, -0.525738f, -0.809012f,  1.0f, 1.0f, 1.0f,  0.863636f, 0.236191f,
-        0.951058f,  0.000000f,  0.309013f,  1.0f, 1.0f, 1.0f,  0.636364f, 0.157461f,
-        0.951058f,  0.000000f, -0.309013f,  1.0f, 1.0f, 1.0f,  0.727273f, 0.157461f,
-        0.000000f,  0.000000f,  1.000000f,  1.0f, 1.0f, 1.0f,  0.681818f, 0.236191f,
-        0.587786f,  0.000000f,  0.809017f,  1.0f, 1.0f, 1.0f,  0.454546f, 0.157461f,
-        -0.951058f,  0.000000f,  0.309013f,  1.0f, 1.0f, 1.0f,  0.545455f, 0.157461f,
-        -0.587786f,  0.000000f,  0.809017f,  1.0f, 1.0f, 1.0f,  0.500000f, 0.236191f,
-        -0.587786f,  0.000000f, -0.809017f,  1.0f, 1.0f, 1.0f,  0.227273f, 0.236191f,
-        -0.951058f,  0.000000f, -0.309013f,  1.0f, 1.0f, 1.0f,  0.045455f, 0.236191f,
-        0.587786f,  0.000000f, -0.809017f,  1.0f, 1.0f, 1.0f,  0.772727f, 0.236191f,
-        0.000000f,  0.000000f, -1.000000f,  1.0f, 1.0f, 1.0f,  0.590909f, 0.236191f,
-        0.688189f,  0.525736f,  0.499997f,  1.0f, 1.0f, 1.0f,  0.409092f, 0.236191f,
-        -0.262869f,  0.525738f,  0.809012f,  1.0f, 1.0f, 1.0f,  0.181819f, 0.314921f,
-        -0.850648f,  0.525736f,  0.000000f,  1.0f, 1.0f, 1.0f,  0.272728f, 0.314921f,
-        -0.262869f,  0.525738f, -0.809012f,  1.0f, 1.0f, 1.0f,  0.227273f, 0.393651f,
-        0.688189f,  0.525736f, -0.499997f,  1.0f, 1.0f, 1.0f,  0.000000f, 0.314921f,
-        0.162456f,  0.850654f,  0.499995f,  1.0f, 1.0f, 1.0f,  0.090910f, 0.314921f,
-        0.525730f,  0.850652f,  0.000000f,  1.0f, 1.0f, 1.0f,  0.045455f, 0.393651f,
-        -0.425323f,  0.850654f,  0.309011f,  1.0f, 1.0f, 1.0f,  0.727273f, 0.314921f,
-        -0.425323f,  0.850654f, -0.309011f,  1.0f, 1.0f, 1.0f,  0.818182f, 0.314921f,
-        0.162456f,  0.850654f, -0.499995f,  1.0f, 1.0f, 1.0f,  0.772727f, 0.393651f
-    };
-
-    static int[] indices = new int[] {
-        0, 13, 12, 1, 13, 15, 0, 12, 18, 0, 18, 20, 0, 20, 17, 1, 15, 22, 2, 14, 24, 3, 19, 26, 4, 21, 28, 5, 22, 31,
-        1, 22, 25, 2, 24, 27, 3, 26, 30, 4, 28, 32, 5, 31, 23, 6, 33, 38, 7, 34, 40, 8, 35, 41, 9, 36, 42, 10, 37, 39,
-        39, 42, 11, 39, 37, 42, 37, 10, 42, 42, 41, 11, 42, 36, 41, 36, 9, 41, 41, 40, 11, 41, 35, 40, 35, 8, 40, 40, 38, 11,
-        40, 34, 38, 34, 7, 38, 38, 39, 11, 38, 33, 39, 33, 6, 39, 23, 37, 11, 23, 31, 37, 31, 10, 37, 32, 36, 10, 32, 29, 36,
-        29, 9, 36, 30, 35, 9, 30, 27, 35, 27, 8, 35, 28, 34, 8, 28, 25, 34, 25, 7, 34, 26, 33, 7, 26, 23, 33, 23, 6, 33,
-        31, 32, 10, 31, 22, 32, 22, 5, 32, 29, 30, 9, 29, 21, 30, 21, 4, 30, 27, 28, 8, 27, 24, 28, 24, 3, 28, 25, 26, 7,
-        25, 22, 26, 22, 2, 26, 23, 24, 11, 23, 16, 24, 16, 6, 24, 17, 22, 6, 17, 20, 22, 20, 5, 22, 20, 21, 5, 20, 18, 21,
-        18, 4, 21, 18, 19, 4, 18, 13, 19, 13, 3, 19, 16, 17, 6, 16, 14, 17, 14, 1, 17, 13, 15, 3, 13, 14, 15, 14, 2, 15
-    };
-
-    public static Mesh createSphereAndScale(float scale) {
-        float[] scaledVertices = vertices.clone();
-
-        for (int i = 0; i < scaledVertices.length; i+=8) {
-            scaledVertices[i] *= scale;
-            scaledVertices[i + 1] *= scale;
-            scaledVertices[i + 2] *= scale;
-        }
-        
-        return Mesh.createMeshWithColor(scaledVertices, indices);
-    }
-    
-}
