+++ src/main/java/com/steve/utils/FileUtil.java
@@ -3,7 +3,6 @@ package com.steve.utils;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Paths;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.regex.Pattern;
