--- src/main/resources/cube.mtl
@@ -1,2 +1,0 @@
-# Blender 4.3.2 MTL File: 'None'
-# www.blender.org
