+++ src/main/resources/models/David head.mtl
@@ -1,0 +1,2 @@
+# Blender 4.3.2 MTL File: 'None'
+# www.blender.org
