+++ src/main/resources/models/cube.obj
@@ -1,0 +1,73 @@
+# Blender 4.3.2
+# www.blender.org
+mtllib cube.mtl
+o Cube
+v -1.000000 -1.000000 1.000000
+v 1.000000 -1.000000 1.000000
+v 1.000000 -1.000000 -1.000000
+v 1.000000 -1.000000 -1.000000
+v 1.000000 1.000000 -1.000000
+v -1.000000 1.000000 -1.000000
+v 1.000000 -1.000000 -1.000000
+v 1.000000 1.000000 -1.000000
+v -1.000000 -1.000000 -1.000000
+v -1.000000 1.000000 -1.000000
+v 1.000000 -1.000000 -1.000000
+v 1.000000 1.000000 -1.000000
+v 1.000000 -1.000000 -1.000000
+v 1.000000 -1.000000 1.000000
+v 1.000000 1.000000 -1.000000
+v 1.000000 1.000000 1.000000
+v 1.000000 -1.000000 1.000000
+v 1.000000 1.000000 1.000000
+v 1.000000 -1.000000 1.000000
+v -1.000000 -1.000000 1.000000
+v 1.000000 1.000000 1.000000
+v -1.000000 1.000000 1.000000
+v -1.000000 -1.000000 1.000000
+v -1.000000 -1.000000 -1.000000
+v -1.000000 1.000000 1.000000
+v -1.000000 -1.000000 1.000000
+v -1.000000 1.000000 -1.000000
+v -1.000000 -1.000000 -1.000000
+v -1.000000 1.000000 1.000000
+v 1.000000 1.000000 1.000000
+v -1.000000 1.000000 -1.000000
+v -1.000000 1.000000 1.000000
+v 1.000000 1.000000 1.000000
+v -1.000000 1.000000 -1.000000
+v -1.000000 -1.000000 -1.000000
+v -1.000000 -1.000000 1.000000
+vn -0.0000 1.0000 -0.0000
+vn -0.0000 -1.0000 -0.0000
+vn -1.0000 -0.0000 -0.0000
+vn -0.0000 -0.0000 1.0000
+vn 1.0000 -0.0000 -0.0000
+vn -0.0000 -0.0000 -1.0000
+vt 0.875000 0.500000
+vt 0.625000 0.750000
+vt 0.625000 0.500000
+vt 0.375000 0.500000
+vt 0.125000 0.750000
+vt 0.125000 0.500000
+vt 0.625000 0.000000
+vt 0.375000 0.250000
+vt 0.375000 0.000000
+vt 0.375000 1.000000
+vt 0.375000 0.750000
+vt 0.875000 0.750000
+vt 0.625000 1.000000
+vt 0.625000 0.250000
+s 0
+f 34/1/1 33/2/1 5/3/1
+f 3/4/2 36/5/2 35/6/2
+f 25/7/3 24/8/3 26/9/3
+f 18/2/4 20/10/4 19/11/4
+f 31/1/1 32/12/1 30/2/1
+f 4/4/2 2/11/2 1/5/2
+f 21/2/4 22/13/4 23/10/4
+f 15/3/5 16/2/5 17/11/5
+f 6/14/6 8/3/6 7/4/6
+f 29/7/3 27/14/3 28/8/3
+f 10/14/6 11/4/6 9/8/6
+f 12/3/5 14/11/5 13/4/5
