+++ src/main/resources/models/objTest.mtl
@@ -1,0 +1,22 @@
+# Blender 4.3.2 MTL File: 'None'
+# www.blender.org
+
+newmtl 材质
+Ns 25.799351
+Ka 1.000000 1.000000 1.000000
+Kd 0.553031 0.800424 0.414418
+Ks 0.500000 0.500000 0.500000
+Ke 0.000000 0.000000 0.000000
+Ni 952.399963
+d 1.000000
+illum 3
+
+newmtl 材质.001
+Ns 1000.000000
+Ka 1.000000 1.000000 1.000000
+Kd 0.800149 0.084692 0.225372
+Ks 0.500000 0.500000 0.500000
+Ke 0.000000 0.000000 0.000000
+Ni 55.699997
+d 1.000000
+illum 3
