+++ src/main/resources/models/objTest.obj
@@ -1,0 +1,145 @@
+# Blender 4.3.2
+# www.blender.org
+mtllib objTest.mtl
+o 立方体
+v -1.000000 -1.000000 1.000000
+v -1.000000 1.000000 1.000000
+v -1.000000 -1.000000 -1.000000
+v -1.000000 1.000000 -1.000000
+v 1.000000 -1.000000 1.000000
+v 1.000000 1.000000 1.000000
+v 1.000000 -1.000000 -1.000000
+v 1.000000 1.000000 -1.000000
+v -1.000000 -1.000000 -1.000000
+v -1.000000 -1.000000 -1.000000
+v -1.000000 -1.000000 -1.000000
+v -1.000000 -1.000000 1.000000
+v -1.000000 -1.000000 1.000000
+v -1.000000 -1.000000 1.000000
+v -1.000000 -1.000000 1.000000
+v -1.000000 1.000000 1.000000
+v -1.000000 1.000000 1.000000
+v -1.000000 1.000000 1.000000
+v -1.000000 1.000000 -1.000000
+v -1.000000 1.000000 -1.000000
+v -1.000000 1.000000 -1.000000
+v -1.000000 1.000000 -1.000000
+v 1.000000 -1.000000 -1.000000
+v 1.000000 -1.000000 -1.000000
+v 1.000000 -1.000000 -1.000000
+v 1.000000 -1.000000 -1.000000
+v 1.000000 1.000000 -1.000000
+v 1.000000 1.000000 -1.000000
+v 1.000000 1.000000 -1.000000
+v 1.000000 -1.000000 1.000000
+v 1.000000 -1.000000 1.000000
+v 1.000000 -1.000000 1.000000
+v 1.000000 1.000000 1.000000
+v 1.000000 1.000000 1.000000
+v 1.000000 1.000000 1.000000
+v 1.000000 1.000000 1.000000
+vn -1.0000 -0.0000 -0.0000
+vn -0.0000 -0.0000 -1.0000
+vn 1.0000 -0.0000 -0.0000
+vn -0.0000 -0.0000 1.0000
+vn -0.0000 -1.0000 -0.0000
+vn -0.0000 1.0000 -0.0000
+vt 0.625000 0.000000
+vt 0.375000 0.250000
+vt 0.375000 0.000000
+vt 0.625000 0.250000
+vt 0.375000 0.500000
+vt 0.625000 0.500000
+vt 0.375000 0.750000
+vt 0.625000 0.750000
+vt 0.375000 1.000000
+vt 0.125000 0.750000
+vt 0.125000 0.500000
+vt 0.875000 0.500000
+vt 0.625000 1.000000
+vt 0.875000 0.750000
+s 0
+usemtl 材质
+f 16/1/1 9/2/1 12/3/1
+f 22/4/2 23/5/2 11/2/2
+f 8/6/3 30/7/3 25/5/3
+f 36/8/4 13/9/4 32/7/4
+f 7/5/5 1/10/5 3/11/5
+f 4/12/6 6/8/6 29/6/6
+f 17/1/1 19/4/1 10/2/1
+f 20/4/2 27/6/2 24/5/2
+f 28/6/3 33/8/3 31/7/3
+f 34/8/4 18/13/4 14/9/4
+f 26/5/5 5/7/5 15/10/5
+f 21/12/6 2/14/6 35/8/6
+o 立方体.001
+v -1.000000 -1.000000 -2.858977
+v -1.000000 1.000000 -2.858977
+v -1.000000 -1.000000 -4.858976
+v -1.000000 1.000000 -4.858976
+v 1.000000 -1.000000 -2.858977
+v 1.000000 1.000000 -2.858977
+v 1.000000 -1.000000 -4.858976
+v 1.000000 1.000000 -4.858976
+v -1.000000 -1.000000 -4.858976
+v -1.000000 -1.000000 -4.858976
+v -1.000000 -1.000000 -4.858976
+v -1.000000 -1.000000 -2.858977
+v -1.000000 -1.000000 -2.858977
+v -1.000000 -1.000000 -2.858977
+v -1.000000 -1.000000 -2.858977
+v -1.000000 1.000000 -2.858977
+v -1.000000 1.000000 -2.858977
+v -1.000000 1.000000 -2.858977
+v -1.000000 1.000000 -4.858976
+v -1.000000 1.000000 -4.858976
+v -1.000000 1.000000 -4.858976
+v -1.000000 1.000000 -4.858976
+v 1.000000 -1.000000 -4.858976
+v 1.000000 -1.000000 -4.858976
+v 1.000000 -1.000000 -4.858976
+v 1.000000 -1.000000 -4.858976
+v 1.000000 1.000000 -4.858976
+v 1.000000 1.000000 -4.858976
+v 1.000000 1.000000 -4.858976
+v 1.000000 -1.000000 -2.858977
+v 1.000000 -1.000000 -2.858977
+v 1.000000 -1.000000 -2.858977
+v 1.000000 1.000000 -2.858977
+v 1.000000 1.000000 -2.858977
+v 1.000000 1.000000 -2.858977
+v 1.000000 1.000000 -2.858977
+vn -1.0000 -0.0000 -0.0000
+vn -0.0000 -0.0000 -1.0000
+vn 1.0000 -0.0000 -0.0000
+vn -0.0000 -0.0000 1.0000
+vn -0.0000 -1.0000 -0.0000
+vn -0.0000 1.0000 -0.0000
+vt 0.625000 0.000000
+vt 0.375000 0.250000
+vt 0.375000 0.000000
+vt 0.625000 0.250000
+vt 0.375000 0.500000
+vt 0.625000 0.500000
+vt 0.375000 0.750000
+vt 0.625000 0.750000
+vt 0.375000 1.000000
+vt 0.125000 0.750000
+vt 0.125000 0.500000
+vt 0.875000 0.500000
+vt 0.625000 1.000000
+vt 0.875000 0.750000
+s 0
+usemtl 材质.001
+f 52/15/7 45/16/7 48/17/7
+f 58/18/8 59/19/8 47/16/8
+f 44/20/9 66/21/9 61/19/9
+f 72/22/10 49/23/10 68/21/10
+f 43/19/11 37/24/11 39/25/11
+f 40/26/12 42/22/12 65/20/12
+f 53/15/7 55/18/7 46/16/7
+f 56/18/8 63/20/8 60/19/8
+f 64/20/9 69/22/9 67/21/9
+f 70/22/10 54/27/10 50/23/10
+f 62/19/11 41/21/11 51/24/11
+f 57/26/12 38/28/12 71/22/12
