+++ src/main/resources/models/ttqobj.mtl
@@ -1,0 +1,2 @@
+# Blender 3.6.4 MTL File: 'None'
+# www.blender.org
