+++ src/main/resources/shaders/f2.fs
@@ -41,20 +41,24 @@ void main()
     vec3 lightDir = normalize(light.position - fragPos);
     float diff = max(dot(norm, lightDir), 0.0);
 
-    vec3 viewDir = normalize(viewPos - fragPos);
-    vec3 reflectDir = reflect(-lightDir, norm);
-    float spec = pow(max(dot(viewDir, reflectDir), 0.0), material.shininess);
+    // vec3 viewDir = normalize(viewPos - fragPos);
+    // vec3 reflectDir = reflect(-lightDir, norm);
+    // float spec = pow(max(dot(viewDir, reflectDir), 0.0), material.shininess);
 
-    vec3 ambient = light.ambient * texture(material.diffuse, texCoord).rgb;
-    vec3 diffuse = light.diffuse * diff * texture(material.diffuse, texCoord).rgb;
-    vec3 specular = light.specular * spec * texture(material.specular, texCoord).rgb;
-    vec3 emission = texture(material.emission, vec2(texCoord.x, texCoord.y + move_time)).rgb;
+    vec3 objectColor = vec3(1.0, 1.0, 1.0);
+    // vec3 objectColor = texture(material.diffuse, texCoord).rgb
+    // vec3 specularText = texture(material.specular, texCoord).rgb
 
-    FragColor = vec4(ambient + diffuse + specular + emission, 1.0);    
+    vec3 ambient = light.ambient * objectColor;
+    vec3 diffuse = light.diffuse * diff * objectColor;
+    // vec3 specular = light.specular * spec * texture(material.specular, texCoord).rgb;
+    // vec3 emission = texture(material.emission, vec2(texCoord.x, texCoord.y + move_time)).rgb;
+
+    FragColor = vec4(ambient + diffuse, 1.0);    
     // FragColor = texture(material.emission, texCoord);
     // FragColor = vec4(vec3(gl_FragCoord.z), 1.0);
 
     // float depth = LinearizeDepth(gl_FragCoord.z) / far; // 为了演示除以 far
-    // FragColor = vec4(vec3(depth), 1.0);
+    // FragColor = vec4((vec3(0.5) - vec3(depth)), 1.0);
 
 }
