--- src/main/resources/ttqobj.mtl
@@ -1,2 +1,0 @@
-# Blender 3.6.4 MTL File: 'None'
-# www.blender.org
